requirements.txt
aiosqlite
py-cord
python-dotenv
requests
boto3
